feat: remove action for copying quick links
Removes the action in favour of the new action for copying permanent links.
