+++ packages/web-pkg/src/composables/actions/files/index.ts
@@ -2,7 +2,7 @@ export * from './useFileActions'
 export * from './useFileActionsEnableSync'
 export * from './useFileActionsToggleHideShare'
 export * from './useFileActionsCopy'
-export * from './useFileActionsCopyQuicklink'
+export * from './useFileActionsCopyPermanentLink'
 export * from './useFileActionsDisableSync'
 export * from './useFileActionsDelete'
 export * from './useFileActionsDownloadArchive'
@@ -25,4 +25,3 @@ export * from './useFileActionsCreateNewShortcut'
 export * from './useFileActionsOpenShortcut'
 export * from './useFileActionsCreateLink'
 export * from './useFileActionsOpenWithApp'
-export * from './useFileActionsCopyPermanentLink'
